# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "InvoiceResponseDto": {
        "properties": {
          "approvedAmountCents": {
            "example": 50000,
            "maximum": 2147483647,
            "minimum": 0,
            "type": "number"
          },
          "competence": {
            "example": "2026-07",
            "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
            "type": "string"
          },
          "contractId": {
            "format": "uuid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "dueDate": {
            "example": "2026-07-10",
            "format": "date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "outstandingAmountCents": {
            "example": 100000,
            "maximum": 2147483647,
            "minimum": 0,
            "type": "number"
          },
          "payments": {
            "items": {
              "$ref": "#/components/schemas/PaymentResponseDto"
            },
            "type": "array"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/InvoiceStatus"
              }
            ]
          },
          "totalValueCents": {
            "example": 150000,
            "maximum": 2147483647,
            "minimum": 1,
            "type": "number"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "contractId",
          "competence",
          "totalValueCents",
          "approvedAmountCents",
          "outstandingAmountCents",
          "dueDate",
          "status",
          "payments",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "InvoiceStatus": {
        "enum": [
          "OPEN",
          "UNDER_REVIEW",
          "PARTIALLY_PAID",
          "PAID",
          "OVERDUE"
        ],
        "type": "string"
      },
...
      "PaginatedInvoicesResponseDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/InvoiceResponseDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "ProblemDetailsDto": {
        "properties": {
          "detail": {
            "example": "A requisição contém dados inválidos.",
            "type": "string"
          },
          "errors": {
            "description": "Demais mensagens de validação, quando houver mais de uma.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "instance": {
            "example": "/contracts?page=0",
            "type": "string"
          },
          "requestId": {
            "example": "0f2a9a56-b0d5-4ac5-8f1e-791e9fc92347",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "example": 400,
            "maximum": 599,
            "minimum": 400,
            "type": "number"
          },
          "timestamp": {
            "example": "2026-07-12T18:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "title": {
            "example": "BadRequestException",
            "type": "string"
          },
          "type": {
            "example": "https://tenancy-ledger.local/problems/http-400",
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "requestId",
          "timestamp"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "JWT de acesso.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "API para gestão auditável de inquilinos, imóveis, contratos e cobranças.",
    "title": "Tenancy Ledger API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/invoices": {
      "get": {
        "operationId": "BillingController_list",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "contractId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "competence",
            "required": false,
            "schema": {
              "example": "2026-07",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/InvoiceStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedInvoicesResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar faturas",
        "tags": [
          "Faturas e pagamentos"
        ]
      }
    },
    "/invoices/{id}": {
      "get": {
        "operationId": "BillingController_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Fatura não encontrada.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar fatura e seus pagamentos",
        "tags": [
          "Faturas e pagamentos"
        ]
      }
    },
    "/invoices/{id}/payments": {
      "post": {
        "operationId": "BillingController_submitPayment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Chave opaca e única por fatura (8 a 128 caracteres ASCII visíveis).",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[\\x21-\\x7e]{8,128}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SubmitPaymentMultipartDto"
              }
            }
          },
          "description": "Comprovante obrigatório para PIX/transferência e proibido para dinheiro.",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Fatura não encontrada.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Chave de idempotência reutilizada ou estado conflitante da fatura.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Regra de negócio ou valor de domínio inválido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submeter pagamento para revisão",
        "tags": [
          "Faturas e pagamentos"
        ]
      }
    },
    "/invoices/{invoiceId}/payments/{paymentId}/approve": {
      "patch": {
        "operationId": "BillingController_approvePayment",
        "parameters": [
          {
            "in": "path",
            "name": "invoiceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "paymentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Fatura não encontrada.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Pagamento já revisado ou aprovação inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aprovar pagamento submetido",
        "tags": [
          "Faturas e pagamentos"
        ]
      }
    },
...
  },
  "servers": [],
...
}